+++ mod.ts
@@ -1,3 +1,4 @@
 export { ArrayDeque } from "./src/ArrayDeque.ts";
 export { LinkedListDeque } from "./src/LinkedListDeque.ts";
 export { ArrayHeapPQ } from "./src/ArrayHeapPQ.ts";
+export { KDTree } from "./src/KDTree.ts";