+++ CHANGELOG.md
@@ -5,12 +5,22 @@ Important changes are documented here. This project aims to follow the
 
 ## 0.x.x
 
+### 0.2.x
+
+#### 0.2.0
+
+- Refactored some variables
+- Implemented the k-d tree data structure (class `KDTree`)
+  - Implemented a reference (a.k.a. slow) k-d tree for testing purposes
+  - Implemented and passed tests
+- Renamed test names so they reflect which data structure they test
+
 ### 0.1.x
 
 #### 0.1.0
 
 - Initial release
 - Implemented and tested the following data structures:
-  - Array deque
-  - Linked list deque
-  - Array-heap extrinsic priority queue
+  - Array deque (class `ArrayDeque`)
+  - Linked list deque (class `LinkedListDeque`)
+  - Array-heap extrinsic priority queue (class `ArrayHeapPQ`)